[WIP] Expandable blocks in allocator
Common advice we give for handling memory fragmentation issues is to
allocate a big block upfront to reserve memory which will get split up later.
For programs with changing tensor sizes this can be especially helpful to
avoid OOMs that happen the first time we see a new largest input and would
otherwise have to allocate new segments.

However the issue with allocating a block upfront is that is nearly impossible
to correctly estimate the size of that block. If too small, space in the block
will run out and the allocator will allocate separate blocks anyway. Too large,
and other non-PyTorch libraries might stop working because they cannot allocate
any memory.

This patch provides the same benefits as using a pre-allocating block but
without having to choose its size upfront. Using the cuMemMap-style APIs,
it adds the ability to expand the last block in a segment when more memory is
needed.

Compared to universally using cudaMallocAsync to avoid fragmentation,
this patch can fix this common fragmentation issue while preserving most
of the existing allocator behavior. Since it is a minimal change to the
allocator, and does not introduce additional global accounting structures,
it is possible to enable/disable expanding blocks at runtime. This should allow users
to, for instance, allocate long-lived parameters and state in individual buffers,
and put temporary state into the large expandable blocks, further reducing
fragmentation.

Limitations:

The primary limitation is that memory can only be freed at the end of of an
expanded block. If there are "holes" in the middle of the block, those
cannot be freed. Future patches can add this ability but the accounting
for it will increase the complexity of the change.

TODO:
* This behavior will be put under an allocator flag, enabled for now
  to see what breaks across all tests.
* Addition to test suite to make sure we hit all new places in the code.

ghstack-source-id: 7e4ee79d79941a972abe44d4a4c7eef6a3a8c6bc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96995